+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Smart contract: loading contract to JVM
 - Smart contract: simple contact validation
 
+## Changed
+- Core: chain synchronization improved (request for chain of blocks instead of last block only)
+- Core: time synchronization improved (based on time-synchronized nodes)
+
 
 ## [1.3.0] - 2018-11-16
 ## Added